refactor!: style and move simple governance items to module
